+++ src/utilities/CacheUtility.php
@@ -7,8 +7,6 @@ namespace putyourlightson\blitz\utilities;
 
 use Craft;
 use craft\base\Utility;
-use craft\helpers\FileHelper;
-use putyourlightson\blitz\Blitz;
 use putyourlightson\blitz\records\CacheRecord;
 
 class CacheUtility extends Utility